Move type declarations to separate file
Created a new folder 'types' and moved all type declarations to 'PostTypes.tsx'
