+++ src/data/MockData.tsx
@@ -1,24 +1,5 @@
 import { faker } from "@faker-js/faker";
-
-type Comment = {
-  username: string;
-  commentText: string;
-};
-
-export type User = {
-  username: string;
-  profilePic: string;
-};
-
-export type Post = {
-  user: User;
-  location: string;
-  image: string;
-  caption: string;
-  likes: User[];
-  comments: Comment[];
-  timestamp: Date;
-};
+import { Post } from "../types/PostTypes";
 
 const generatePostData = (): Post => {
   return {
