+++ src/types/PostTypes.tsx
@@ -1,0 +1,19 @@
+type Comment = {
+  username: string;
+  commentText: string;
+};
+
+export type User = {
+  username: string;
+  profilePic: string;
+};
+
+export type Post = {
+  user: User;
+  location: string;
+  image: string;
+  caption: string;
+  likes: User[];
+  comments: Comment[];
+  timestamp: Date;
+};
